+++ app/__main__.py
@@ -4,7 +4,7 @@
 from os import environ as env
 from sys import exit
 
-# TODO conditional imports and env best practices
+# TODO conditional imports and env best practices, maybe use parsed .env with dotenv
 # conditional to avoid flake8 E402 module level import not at top of file
 if True:
     env["APP_DEBUG_IS_ON"] = "True"